adds paginate and subscription update

+++ services/contact.js
@@ -1,23 +1,49 @@
 const { Contact } = require('../models')
 
-const getAll = () => {
-  return Contact.find({})
+const getAll = async (userId, { page = 1, limit = 20, favorite }) => {
+  const filter = favorite ? { favorite: `${favorite}` } : null
+
+  const { docs: contacts, totalDocs: total } = await Contact.paginate(
+    {
+      ...filter,
+      owner: userId,
+    },
+    {
+      page,
+      limit,
+      populate: {
+        path: 'owner',
+        select: ' email subscription',
+      },
+    }
+  )
+
+  return { contacts, total, page: Number(page), limit: Number(limit) }
 }
 
-const findById = (id) => {
-  return Contact.findById(id)
+const findById = async (userId, id) => {
+  return await Contact.findById({ _id: id, owner: userId }).populate({
+    path: 'owner',
+    select: ' email subscription',
+  })
 }
 
-const add = (newContact) => {
-  return Contact.create(newContact)
+const add = async (userId, newContact) => {
+  return await Contact.create({ ...newContact, owner: userId })
 }
 
-const update = (id, updateContact) => {
-  return Contact.findByIdAndUpdate(id, updateContact, { new: true })
+const update = async (userId, id, updateContact) => {
+  return await Contact.findByIdAndUpdate(
+    { _id: id, owner: userId },
+    updateContact,
+    {
+      new: true,
+    }
+  )
 }
 
-const remove = (id) => {
-  return Contact.findByIdAndRemove(id)
+const remove = async (userId, id) => {
+  return await Contact.findByIdAndRemove({ _id: id, owner: userId })
 }
 
 module.exports = {
